Move get_savers_for_track
